+++ packages/integrations/vue-devtools/src/config.ts
