+++ packages/integrations/vue-devtools/src/views/index.ts
@@ -1,0 +1,3 @@
+export * from './createInspectorViews'
+export * from './createSettingsViews'
+export * from './createTimelineViews'
