+++ packages/integrations/vue-devtools/src/views/inspector/index.ts
@@ -1,0 +1,2 @@
+export * from './createMainScenarioView'
+export * from './createScenarioDetailView'
